[05] 소스코드/bigdata/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.1
scikit-surprise==1.1.0
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
surprise==0.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.9
